+++ pyrate_limiter/__init__.py
@@ -1,8 +1,2 @@
 # flake8: noqa
-"""PyrateLimiter
-"""
-from .abstracts import *
-from .buckets import *
-from .clocks import *
-from .exceptions import *
-from .limiter import *
+from .limiter import Limiter
